refactor(plugin-workflow): keep context singleton imports minimal
