+++ packages/plugins/@nocobase/plugin-workflow/src/client-v2/canvas/NodeDragContext.tsx
@@ -18,13 +18,13 @@
  * `FlowContext` directly, and the pure graph walks from the relocated modules.
  */
 
-import React, { useCallback, useContext, useEffect, useMemo, useRef, useState } from 'react';
+import React, { createContext, useCallback, useContext, useEffect, useMemo, useRef, useState } from 'react';
 import { App, Checkbox } from 'antd';
 import { useFlowContext as useFlowEngineContext } from '@nocobase/flow-engine';
 
 import { useFlowContext, useWorkflowCanvasExecuted, type CanvasNode } from './contexts';
 import { useT, useWorkflowTranslation } from '../locale';
-import { getWorkflowSingleton } from '../contextSingleton';
+import { getWorkflowSingleton } from '../utils/contextSingleton';
 import useStyles from './style';
 import { useWorkflowPlugin } from './useWorkflowInstruction';
 import type { Instruction } from './Instruction';
@@ -113,9 +113,7 @@ export type NodeDragContextValue = {
   consumeClick: () => boolean;
 };
 
-const NodeDragContext = getWorkflowSingleton('NodeDragContext', () =>
-  React.createContext<NodeDragContextValue | null>(null),
-);
+const NodeDragContext = getWorkflowSingleton('NodeDragContext', () => createContext<NodeDragContextValue | null>(null));
 
 export function useNodeDragContext() {
   return useContext(NodeDragContext);
