+++ packages/plugins/@nocobase/plugin-workflow/src/client-v2/canvas/RemoveNodeContext.tsx
@@ -22,12 +22,12 @@
  * provider and pass its own runtime (the allowed v1 → v2 direction).
  */
 
-import React, { useCallback, useContext, useMemo, useState } from 'react';
+import React, { createContext, useCallback, useContext, useMemo, useState } from 'react';
 import { App, Modal, Radio, Select, Space } from 'antd';
 import { useFlowEngine } from '@nocobase/flow-engine';
 import { useFlowContext } from './contexts';
 import { useT } from '../locale';
-import { getWorkflowSingleton } from '../contextSingleton';
+import { getWorkflowSingleton } from '../utils/contextSingleton';
 import { PluginWorkflowClientV2 } from '../plugin';
 import { collectBranchNodes, findNodesReferencing } from './removeNodeUtils';
 
@@ -62,7 +62,7 @@ function useModernCanvasRuntime(): CanvasRemoveRuntime {
 }
 
 const RemoveNodeContext = getWorkflowSingleton('RemoveNodeContext', () =>
-  React.createContext<RemoveNodeContextValue | null>(null),
+  createContext<RemoveNodeContextValue | null>(null),
 );
 
 export function useRemoveNodeContext() {
