+++ packages/plugins/@nocobase/plugin-workflow/src/client-v2/canvas/contexts.tsx
@@ -23,9 +23,9 @@
  * bundled into the node-context value.
  */
 
-import React, { useContext } from 'react';
+import React, { createContext, useContext } from 'react';
 import type { WorkflowCanvasRecord, WorkflowRevision } from '../components/workflowCanvas';
-import { getWorkflowSingleton } from '../contextSingleton';
+import { getWorkflowSingleton } from '../utils/contextSingleton';
 
 /**
  * A canvas node — the live linked-list element produced by `linkNodes`: the flat
@@ -74,9 +74,7 @@ export type WorkflowCanvasFlowContextValue = {
   userJob?: Record<string, any> | null;
 };
 
-export const FlowContext = getWorkflowSingleton('FlowContext', () =>
-  React.createContext<WorkflowCanvasFlowContextValue>({}),
-);
+export const FlowContext = getWorkflowSingleton('FlowContext', () => createContext<WorkflowCanvasFlowContextValue>({}));
 
 export function useFlowContext() {
   return useContext(FlowContext);
@@ -85,7 +83,7 @@ export function useFlowContext() {
 // Holds the bare workflow record (v1's split from FlowContext). Default `{}` (not null) preserves v1's behavior, so
 // unguarded `.type`/`.config` reads at existing call sites stay safe; `Partial` makes the empty default assignable.
 export const CurrentWorkflowContext = getWorkflowSingleton('CurrentWorkflowContext', () =>
-  React.createContext<Partial<WorkflowCanvasRecord>>({}),
+  createContext<Partial<WorkflowCanvasRecord>>({}),
 );
 
 export function useCurrentWorkflowContext() {
@@ -93,7 +91,7 @@ export function useCurrentWorkflowContext() {
 }
 
 export const WorkflowVariableSourceContext = getWorkflowSingleton('WorkflowVariableSourceContext', () =>
-  React.createContext<Partial<WorkflowCanvasRecord> | null>(null),
+  createContext<Partial<WorkflowCanvasRecord> | null>(null),
 );
 
 export function useWorkflowVariableSourceContext() {
@@ -102,7 +100,7 @@ export function useWorkflowVariableSourceContext() {
 
 // Default `{}` (not null) to match v1's `NodeContext` default, so existing v1 call sites that read
 // `useNodeContext().config` without guarding are unaffected.
-export const NodeContext = getWorkflowSingleton('NodeContext', () => React.createContext<CanvasNode>({} as CanvasNode));
+export const NodeContext = getWorkflowSingleton('NodeContext', () => createContext<CanvasNode>({} as CanvasNode));
 
 export function useNodeContext() {
   return useContext(NodeContext);
